Move the cookie-policy from legal#cookies to path legal/cookie-policy
The cookie policy is independent and required for all users (registered or not), so changes to the cookie policy should not be coupled to the main privacy policy since they're separate responsibilities.
